Refine raft include
