+++ dbms/src/Storages/KVStore/MultiRaft/Spill/RegionUncommittedDataList.h
@@ -14,7 +14,6 @@
 
 #pragma once
 
-#include <Storages/DeltaMerge/ExternalDTFileInfo.h>
 #include <Storages/KVStore/Decode/DecodedTiKVKeyValue.h>
 
 namespace DB
